backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-generativeai==0.8.2
python-dotenv==1.0.1
pydantic==2.9.2
firebase-admin==6.5.0
httpx==0.27.2
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
sendgrid==6.11.0
twilio==8.10.2
email-validator==2.1.0.post1
